refactor(bindings): split the Python and WASM config.rs into modules
The PyO3 and wasm-bindgen config layers were single ~2.4k-line files.
Split each into a `config/` module — parse helpers, refiners, multiscale,
upscale, strategy, and detector submodules — re-exported from
`config/mod.rs`. Pure move: `lib.rs` imports and the generated Python and
JS APIs are byte-identical (pytest 86, the wasm enum-discriminant pin
test, maturin develop, and the wasm32 check all stay green).
